backend/core/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
stripe==7.1.0
